Add an @kind attribute
