+++ src/controls/keycodes.js
@@ -5,6 +5,8 @@ Crafty.extend({
     /**@
      * #Crafty.keys
      * @category Input
+     * @kind Property
+     * 
      * Object of key names and the corresponding Unicode key code.
      *
      * ~~~
@@ -199,6 +201,8 @@ Crafty.extend({
     /**@
      * #Crafty.mouseButtons
      * @category Input
+     * @kind Property
+     * 
      * An object mapping mouseButton names to the corresponding button ID.
      * In all mouseEvents, we add the `e.mouseButton` property with a value normalized to match e.button of modern webkit browsers:
      *
