+++ src/core/scenes.js
@@ -8,6 +8,8 @@ module.exports = {
     /**@
      * #Crafty.scene
      * @category Scenes, Stage
+     * @kind Method
+     * 
      * @trigger SceneChange - just before a new scene is initialized - { oldScene:String, newScene:String }
      * @trigger SceneDestroy - just before the current scene is destroyed - { newScene:String  }
      *
@@ -94,6 +96,7 @@ module.exports = {
     /* 
      * #Crafty.defineScene
      * @category Scenes, Stage
+     * @kind Method
      *
      * @sign public void Crafty.enterScene(String name[, Data])
      * @param name - Name of the scene to run.
@@ -117,6 +120,8 @@ module.exports = {
     /* 
      * #Crafty.enterScene
      * @category Scenes, Stage
+     * @kind Method
+     * 
      * @trigger SceneChange - just before a new scene is initialized - { oldScene:String, newScene:String }
      * @trigger SceneDestroy - just before the current scene is destroyed - { newScene:String  }
      *
